+++ cli/tests/test_operations.rs
@@ -1139,15 +1139,15 @@ fn test_op_diff() {
         .success();
     work_dir.run_jj(["git", "fetch"]).success();
     work_dir
-        .run_jj(["bookmark", "track", "bookmark-1@origin"])
+        .run_jj(["bookmark", "track", "bookmark-1"])
         .success();
 
     // Overview of op log.
     let output = work_dir.run_jj(["op", "log"]);
     insta::assert_snapshot!(output, @r"
-    @  64f6f2454329 test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
+    @  14a49257b1cb test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
     │  track remote bookmark bookmark-1@origin
-    │  args: jj bookmark track bookmark-1@origin
+    │  args: jj bookmark track bookmark-1
     ○  93de354b21ad test-username@host.example.com 2001-02-03 04:05:09.000 +07:00 - 2001-02-03 04:05:09.000 +07:00
     │  fetch from git remote(s) origin
     │  args: jj git fetch
@@ -1169,8 +1169,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff", "--from", "@", "--to", "@"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 64f6f2454329 (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
-      To operation: 64f6f2454329 (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
+    From operation: 14a49257b1cb (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
+      To operation: 14a49257b1cb (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
     [EOF]
     ");
 
@@ -1180,7 +1180,7 @@ fn test_op_diff() {
     let output = work_dir.run_jj(["op", "diff", "--from", "@-", "--to", "@"]);
     insta::assert_snapshot!(output, @r"
     From operation: 93de354b21ad (2001-02-03 08:05:09) fetch from git remote(s) origin
-      To operation: 64f6f2454329 (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
+      To operation: 14a49257b1cb (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
 
     Changed local bookmarks:
     bookmark-1:
@@ -1200,7 +1200,7 @@ fn test_op_diff() {
     let output = work_dir.run_jj(["op", "diff", "--from", "0000000"]);
     insta::assert_snapshot!(output, @r"
     From operation: 000000000000 root()
-      To operation: 64f6f2454329 (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
+      To operation: 14a49257b1cb (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
 
     Changed commits:
     ○  + rnnslrkn 4ff62539 bookmark-2@origin | Commit 2
@@ -1233,7 +1233,7 @@ fn test_op_diff() {
     // Diff from latest operation to root operation
     let output = work_dir.run_jj(["op", "diff", "--to", "0000000"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 64f6f2454329 (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
+    From operation: 14a49257b1cb (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
       To operation: 000000000000 root()
 
     Changed commits:
@@ -1290,12 +1290,12 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "log"]);
     insta::assert_snapshot!(output, @r"
-    @    4ee6d5cabe2e test-username@host.example.com 2001-02-03 04:05:19.000 +07:00 - 2001-02-03 04:05:19.000 +07:00
+    @    8aa8ee21be82 test-username@host.example.com 2001-02-03 04:05:19.000 +07:00 - 2001-02-03 04:05:19.000 +07:00
     ├─╮  reconcile divergent operations
     │ │  args: jj log
-    ○ │  64f6f2454329 test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
+    ○ │  14a49257b1cb test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
     │ │  track remote bookmark bookmark-1@origin
-    │ │  args: jj bookmark track bookmark-1@origin
+    │ │  args: jj bookmark track bookmark-1
     │ ○  e932e147b07d test-username@host.example.com 2001-02-03 04:05:18.000 +07:00 - 2001-02-03 04:05:18.000 +07:00
     ├─╯  point bookmark bookmark-1 to commit 4ff6253913375c6ebdddd8423c11df3b3f17e331
     │    args: jj bookmark set bookmark-1 -r bookmark-2@origin --at-op @-
@@ -1318,8 +1318,8 @@ fn test_op_diff() {
     // Diff between the first parent of the merge operation and the merge operation.
     let output = work_dir.run_jj(["op", "diff", "--from", first_parent_id, "--to", op_id]);
     insta::assert_snapshot!(output, @r"
-    From operation: 64f6f2454329 (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
-      To operation: 4ee6d5cabe2e (2001-02-03 08:05:19) reconcile divergent operations
+    From operation: 14a49257b1cb (2001-02-03 08:05:10) track remote bookmark bookmark-1@origin
+      To operation: 8aa8ee21be82 (2001-02-03 08:05:19) reconcile divergent operations
 
     Changed local bookmarks:
     bookmark-1:
@@ -1334,7 +1334,7 @@ fn test_op_diff() {
     let output = work_dir.run_jj(["op", "diff", "--from", second_parent_id, "--to", op_id]);
     insta::assert_snapshot!(output, @r"
     From operation: e932e147b07d (2001-02-03 08:05:18) point bookmark bookmark-1 to commit 4ff6253913375c6ebdddd8423c11df3b3f17e331
-      To operation: 4ee6d5cabe2e (2001-02-03 08:05:19) reconcile divergent operations
+      To operation: 8aa8ee21be82 (2001-02-03 08:05:19) reconcile divergent operations
 
     Changed local bookmarks:
     bookmark-1:
@@ -1362,8 +1362,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 4ee6d5cabe2e (2001-02-03 08:05:19) reconcile divergent operations
-      To operation: 37661c38b5a1 (2001-02-03 08:05:23) fetch from git remote(s) origin
+    From operation: 8aa8ee21be82 (2001-02-03 08:05:19) reconcile divergent operations
+      To operation: 26d9ccc41b21 (2001-02-03 08:05:23) fetch from git remote(s) origin
 
     Changed commits:
     ○  + kulxwnxm e1a239a5 bookmark-2@origin | Commit 5
@@ -1405,8 +1405,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 37661c38b5a1 (2001-02-03 08:05:23) fetch from git remote(s) origin
-      To operation: 83bf9eca7d31 (2001-02-03 08:05:25) create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
+    From operation: 26d9ccc41b21 (2001-02-03 08:05:23) fetch from git remote(s) origin
+      To operation: 68ab6999295f (2001-02-03 08:05:25) create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
 
     Changed local bookmarks:
     bookmark-2:
@@ -1416,7 +1416,7 @@ fn test_op_diff() {
     ");
 
     // Test tracking of bookmark.
-    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2@origin"]);
+    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
     Started tracking 1 remote bookmarks.
@@ -1424,8 +1424,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 83bf9eca7d31 (2001-02-03 08:05:25) create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
-      To operation: 818db089870a (2001-02-03 08:05:27) track remote bookmark bookmark-2@origin
+    From operation: 68ab6999295f (2001-02-03 08:05:25) create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
+      To operation: dfa9381f7159 (2001-02-03 08:05:27) track remote bookmark bookmark-2@origin
 
     Changed remote bookmarks:
     bookmark-2@origin:
@@ -1436,7 +1436,7 @@ fn test_op_diff() {
 
     // Test creation of new commit.
     // Test tracking of bookmark.
-    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2@origin"]);
+    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
     Warning: Remote bookmark already tracked: bookmark-2@origin
@@ -1445,8 +1445,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 83bf9eca7d31 (2001-02-03 08:05:25) create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
-      To operation: 818db089870a (2001-02-03 08:05:27) track remote bookmark bookmark-2@origin
+    From operation: 68ab6999295f (2001-02-03 08:05:25) create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
+      To operation: dfa9381f7159 (2001-02-03 08:05:27) track remote bookmark bookmark-2@origin
 
     Changed remote bookmarks:
     bookmark-2@origin:
@@ -1466,8 +1466,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 818db089870a (2001-02-03 08:05:27) track remote bookmark bookmark-2@origin
-      To operation: 263f1b7d2e97 (2001-02-03 08:05:31) new empty commit
+    From operation: dfa9381f7159 (2001-02-03 08:05:27) track remote bookmark bookmark-2@origin
+      To operation: 8cdaab259ed6 (2001-02-03 08:05:31) new empty commit
 
     Changed commits:
     ○  + xlzxqlsl 731ab199 (empty) new commit
@@ -1488,8 +1488,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 263f1b7d2e97 (2001-02-03 08:05:31) new empty commit
-      To operation: a90be14f7026 (2001-02-03 08:05:33) point bookmark bookmark-1 to commit 731ab19950fc6fc1199b9ea73cb8b9016f22e8f3
+    From operation: 8cdaab259ed6 (2001-02-03 08:05:31) new empty commit
+      To operation: ea9f7f618485 (2001-02-03 08:05:33) point bookmark bookmark-1 to commit 731ab19950fc6fc1199b9ea73cb8b9016f22e8f3
 
     Changed local bookmarks:
     bookmark-1:
@@ -1508,8 +1508,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: a90be14f7026 (2001-02-03 08:05:33) point bookmark bookmark-1 to commit 731ab19950fc6fc1199b9ea73cb8b9016f22e8f3
-      To operation: 1b2b83317596 (2001-02-03 08:05:35) delete bookmark bookmark-2
+    From operation: ea9f7f618485 (2001-02-03 08:05:33) point bookmark bookmark-1 to commit 731ab19950fc6fc1199b9ea73cb8b9016f22e8f3
+      To operation: fd2ddf76493d (2001-02-03 08:05:35) delete bookmark bookmark-2
 
     Changed local bookmarks:
     bookmark-2:
@@ -1529,8 +1529,8 @@ fn test_op_diff() {
     ");
     let output = work_dir.run_jj(["op", "diff"]);
     insta::assert_snapshot!(output, @r"
-    From operation: 1b2b83317596 (2001-02-03 08:05:35) delete bookmark bookmark-2
-      To operation: b85f801e5384 (2001-02-03 08:05:37) push all tracked bookmarks to git remote origin
+    From operation: fd2ddf76493d (2001-02-03 08:05:35) delete bookmark bookmark-2
+      To operation: 6bfdecc01c75 (2001-02-03 08:05:37) push all tracked bookmarks to git remote origin
 
     Changed remote bookmarks:
     bookmark-1@origin:
@@ -2284,15 +2284,15 @@ fn test_op_show() {
         .success();
     work_dir.run_jj(["git", "fetch"]).success();
     work_dir
-        .run_jj(["bookmark", "track", "bookmark-1@origin"])
+        .run_jj(["bookmark", "track", "bookmark-1"])
         .success();
 
     // Overview of op log.
     let output = work_dir.run_jj(["op", "log"]);
     insta::assert_snapshot!(output, @r"
-    @  64f6f2454329 test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
+    @  14a49257b1cb test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
     │  track remote bookmark bookmark-1@origin
-    │  args: jj bookmark track bookmark-1@origin
+    │  args: jj bookmark track bookmark-1
     ○  93de354b21ad test-username@host.example.com 2001-02-03 04:05:09.000 +07:00 - 2001-02-03 04:05:09.000 +07:00
     │  fetch from git remote(s) origin
     │  args: jj git fetch
@@ -2315,9 +2315,9 @@ fn test_op_show() {
     // Showing the latest operation.
     let output = work_dir.run_jj(["op", "show", "@"]);
     insta::assert_snapshot!(output, @r"
-    64f6f2454329 test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
+    14a49257b1cb test-username@host.example.com 2001-02-03 04:05:10.000 +07:00 - 2001-02-03 04:05:10.000 +07:00
     track remote bookmark bookmark-1@origin
-    args: jj bookmark track bookmark-1@origin
+    args: jj bookmark track bookmark-1
 
     Changed local bookmarks:
     bookmark-1:
@@ -2386,7 +2386,7 @@ fn test_op_show() {
     // Showing a merge operation is empty.
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    f7f71a8bc2f4 test-username@host.example.com 2001-02-03 04:05:17.000 +07:00 - 2001-02-03 04:05:17.000 +07:00
+    d179a8a7f8bf test-username@host.example.com 2001-02-03 04:05:17.000 +07:00 - 2001-02-03 04:05:17.000 +07:00
     reconcile divergent operations
     args: jj log
     [EOF]
@@ -2405,7 +2405,7 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    a32821470ddb test-username@host.example.com 2001-02-03 04:05:19.000 +07:00 - 2001-02-03 04:05:19.000 +07:00
+    dc9a130455e8 test-username@host.example.com 2001-02-03 04:05:19.000 +07:00 - 2001-02-03 04:05:19.000 +07:00
     fetch from git remote(s) origin
     args: jj git fetch
 
@@ -2449,7 +2449,7 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    2ce7ebded2de test-username@host.example.com 2001-02-03 04:05:21.000 +07:00 - 2001-02-03 04:05:21.000 +07:00
+    2e0e79579bb2 test-username@host.example.com 2001-02-03 04:05:21.000 +07:00 - 2001-02-03 04:05:21.000 +07:00
     create bookmark bookmark-2 pointing to commit e1a239a57eb15cefc5910198befbbbe2b43c47af
     args: jj bookmark create bookmark-2 -r bookmark-2@origin
 
@@ -2461,7 +2461,7 @@ fn test_op_show() {
     ");
 
     // Test tracking of a bookmark.
-    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2@origin"]);
+    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
     Started tracking 1 remote bookmarks.
@@ -2469,9 +2469,9 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    9cc95f35ba27 test-username@host.example.com 2001-02-03 04:05:23.000 +07:00 - 2001-02-03 04:05:23.000 +07:00
+    75265616f1b2 test-username@host.example.com 2001-02-03 04:05:23.000 +07:00 - 2001-02-03 04:05:23.000 +07:00
     track remote bookmark bookmark-2@origin
-    args: jj bookmark track bookmark-2@origin
+    args: jj bookmark track bookmark-2
 
     Changed remote bookmarks:
     bookmark-2@origin:
@@ -2481,7 +2481,7 @@ fn test_op_show() {
     ");
 
     // Test creation of new commit.
-    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2@origin"]);
+    let output = work_dir.run_jj(["bookmark", "track", "bookmark-2"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
     Warning: Remote bookmark already tracked: bookmark-2@origin
@@ -2490,9 +2490,9 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    9cc95f35ba27 test-username@host.example.com 2001-02-03 04:05:23.000 +07:00 - 2001-02-03 04:05:23.000 +07:00
+    75265616f1b2 test-username@host.example.com 2001-02-03 04:05:23.000 +07:00 - 2001-02-03 04:05:23.000 +07:00
     track remote bookmark bookmark-2@origin
-    args: jj bookmark track bookmark-2@origin
+    args: jj bookmark track bookmark-2
 
     Changed remote bookmarks:
     bookmark-2@origin:
@@ -2512,7 +2512,7 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    8727dbd86fa3 test-username@host.example.com 2001-02-03 04:05:27.000 +07:00 - 2001-02-03 04:05:27.000 +07:00
+    3117c5ac4af3 test-username@host.example.com 2001-02-03 04:05:27.000 +07:00 - 2001-02-03 04:05:27.000 +07:00
     new empty commit
     args: jj new bookmark-1@origin -m 'new commit'
 
@@ -2535,7 +2535,7 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    f5d053abd6be test-username@host.example.com 2001-02-03 04:05:29.000 +07:00 - 2001-02-03 04:05:29.000 +07:00
+    2e179db04ef9 test-username@host.example.com 2001-02-03 04:05:29.000 +07:00 - 2001-02-03 04:05:29.000 +07:00
     point bookmark bookmark-1 to commit 8f340dd76dc637e4deac17f30056eef7d8eaf682
     args: jj bookmark set bookmark-1 -r @
 
@@ -2556,7 +2556,7 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    97af2adb0e1f test-username@host.example.com 2001-02-03 04:05:31.000 +07:00 - 2001-02-03 04:05:31.000 +07:00
+    7f6cf80e50d5 test-username@host.example.com 2001-02-03 04:05:31.000 +07:00 - 2001-02-03 04:05:31.000 +07:00
     delete bookmark bookmark-2
     args: jj bookmark delete bookmark-2
 
@@ -2578,7 +2578,7 @@ fn test_op_show() {
     ");
     let output = work_dir.run_jj(["op", "show"]);
     insta::assert_snapshot!(output, @r"
-    e2345a3aa1b3 test-username@host.example.com 2001-02-03 04:05:33.000 +07:00 - 2001-02-03 04:05:33.000 +07:00
+    15c5b97a8e12 test-username@host.example.com 2001-02-03 04:05:33.000 +07:00 - 2001-02-03 04:05:33.000 +07:00
     push all tracked bookmarks to git remote origin
     args: jj git push --tracked --deleted
 
@@ -2593,9 +2593,9 @@ fn test_op_show() {
     ");
 
     // Showing a given operation, without graph
-    let output = work_dir.run_jj(["op", "show", "--no-graph", "8727dbd86fa3"]);
+    let output = work_dir.run_jj(["op", "show", "--no-graph", "3117c5ac4af3"]);
     insta::assert_snapshot!(output, @r"
-    8727dbd86fa3 test-username@host.example.com 2001-02-03 04:05:27.000 +07:00 - 2001-02-03 04:05:27.000 +07:00
+    3117c5ac4af3 test-username@host.example.com 2001-02-03 04:05:27.000 +07:00 - 2001-02-03 04:05:27.000 +07:00
     new empty commit
     args: jj new bookmark-1@origin -m 'new commit'